+++ hdsr_fewspy/api_calls/get_locations.py
@@ -57,7 +57,7 @@ class GetLocations(GetRequest):
         if self.output_choice in {OutputChoices.json_response_in_memory, OutputChoices.xml_response_in_memory}:
             return response
 
-        assert self.output_choice == OutputChoices.pandas_dataframe_in_memory, "code error"
+        assert self.output_choice == OutputChoices.pandas_dataframe_in_memory, "code error GetLocations"
         # parse the response to dataframe
         if response.status_code == 200: