+++ hdsr_fewspy/api_calls/time_series/get_time_series_single.py
@@ -56,7 +56,7 @@ class GetTimeSeriesSingle(GetTimeSeriesBase):
         if self.output_choice in {OutputChoices.json_response_in_memory, OutputChoices.xml_response_in_memory}:
             return responses
 
-        assert self.output_choice == OutputChoices.pandas_dataframe_in_memory, "code error"
+        assert self.output_choice == OutputChoices.pandas_dataframe_in_memory, "code error GetTimeSeriesSingle"
         # parse the response to dataframe
         df = response_jsons_to_one_df(
             responses=responses, drop_missing_values=self.drop_missing_values, flag_threshold=self.flag_threshold